+++ scripts/ammunition-system/actions/conjure-bullet.js
@@ -1,9 +1,9 @@
-import { handleReload } from "../../feats/crossbow-feats.js";
 import { PF2eActor } from "../../types/pf2e/actor.js";
 import { PF2eToken } from "../../types/pf2e/token.js";
+import { HookManager } from "../../utils/hook-manager.js";
 import { getControlledActorAndToken, getEffectFromActor, getItem, getItemFromActor, postToChat, setEffectTarget, showWarning, Updates } from "../../utils/utils.js";
 import { getSingleWeapon, getWeapons } from "../../utils/weapon-utils.js";
-import { CONJURED_ROUND_EFFECT_ID, CONJURED_ROUND_ITEM_ID, CONJURE_BULLET_ACTION_ID, CONJURE_BULLET_IMG } from "../constants.js";
+import { CONJURE_BULLET_ACTION_ID, CONJURE_BULLET_IMG, CONJURED_ROUND_EFFECT_ID, CONJURED_ROUND_ITEM_ID } from "../constants.js";
 import { checkFullyLoaded, isFullyLoaded } from "../utils.js";
 import { setLoadedChamber } from "./next-chamber.js";
 
@@ -83,8 +83,9 @@ export async function performConjureBullet(actor, token, weapon) {
         1,
     );
 
-    await handleReload(weapon, updates);
+    await HookManager.call("reload", weapon, updates);
 
     updates.handleUpdates();
+    
     Hooks.callAll("pf2eRangedCombatConjureBullet", actor, token, weapon);
 }