+++ scripts/chat-message-hook.js
@@ -1,3 +1,6 @@
+import { HookManager } from "./utils/hook-manager.js";
+import { Updates } from "./utils/utils.js";
+
 /**
  * When rendering a chat message sent by this module, give it the "hide" class
  * if the current user doesn't have the required permission level
@@ -49,5 +52,47 @@ Hooks.on(
             }
         );
         observer.observe(chatLog[0], { childList: true });
+    }
+);
+
+/**
+ * When we get a message about rolling damage for a weapon, fire the weapon damage hook for that weapon
+ */
+Hooks.on(
+    "preCreateChatMessage",
+    async message => {
+        const actor = message.actor;
+        if (!actor) {
+            return;
+        }
+
+        const flags = message.flags?.pf2e;
+        if (!flags) {
+            return;
+        }
+
+        if (flags.context?.type != "damage-roll") {
+            return;
+        }
+
+        if (flags.origin?.type != "weapon") {
+            return;
+        }
+
+        const uuid = flags.origin?.uuid;
+        if (!uuid) {
+            return;
+        }
+
+        const weapon = await fromUuid(uuid);
+        if (!weapon) {
+            return;
+        }
+
+        const updates = new Updates(weapon.actor);
+
+        await HookManager.call("weapon-damage", weapon, updates);
+
+        updates.handleUpdates();
     }
 );