+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 5.3.0 - 2024-06-06 (Pathfinder 2e 5.16.1)
+### Feature
+- Reintroduced ammunition effects, replacing the system effects, so the correct effect is applied even if ammunition is changed after loading.
+- Update and improve Crossbow Crack Shot automation.
+
 ## 5.2.1 - 2024-06-02 (Pathfinder 2e 5.16.1)
 ### Fixes
 - Fix failing to post messages about using ammunition.